+++ packages/typir-langium/src/features/langium-validation.ts
@@ -5,8 +5,9 @@
  ******************************************************************************/
 
 import { AstNode, LangiumDefaultCoreServices, ValidationAcceptor, ValidationChecks } from 'langium';
-import { TypirServices, ValidationProblem, ValidationRule } from 'typir';
+import { DefaultValidationCollector, TypirServices, ValidationCollector, ValidationProblem, ValidationRule } from 'typir';
 import { LangiumServicesForTypirBinding } from '../typir-langium.js';
+import { LangiumAstTypes } from '../utils/typir-langium-utils.js';
 
 export function registerTypirValidationChecks(langiumServices: LangiumDefaultCoreServices, typirServices: LangiumServicesForTypirBinding) {
     const registry = langiumServices.validation.ValidationRegistry;
@@ -32,13 +33,15 @@ export function registerTypirValidationChecks(langiumServices: LangiumDefaultCor
 *
 * Improved Validation API for Langium:
 * - const ref: (kind: unknown) => kind is FunctionKind = isFunctionKind; // use this signature for Langium?
-* - register validations for AST node $types (similar as Langium does it) => this is much more performant
 * - [<VariableDeclaration>{ selector: isVariableDeclaration, result: languageNode => languageNode.type }, <BinaryExpression>{}]      Array<InferenceRule<T>>
-* - discriminator rule: $type '$VariableDeclaration' + record / "Sprungtabelle" for the Langium-binding (or both in core)? for improved performance (?)
-* - alternativ discriminator rule: unknown => string; AstNode => node.$type; Vorsicht mit Sub-Typen (Vollständigkeit+Updates, no abstract types)!
 * Apply the same ideas for InferenceRules as well!
 */
 
+
+/**
+ * This service is a technical adapter service,
+ * which helps to call the Typir validations, triggered by the Langium validation mechanisms.
+ */
 export interface LangiumTypirValidator {
     /**
      * Will be called once before starting the validation of a LangiumDocument.
@@ -89,4 +92,48 @@ export class DefaultLangiumTypirValidator implements LangiumTypirValidator {
             accept(problem.severity, message, { node, property: problem.languageProperty, index: problem.languageIndex });
         }
     }
+}
+
+
+/**
+ * Taken and adapted from 'ValidationChecks' from 'langium'.
+ *
+ * A utility type for associating non-primitive AST types to corresponding validation rules. For example:
+ *
+ * ```ts
+ *   addValidationsRulesForAstNodes<LoxAstType>({
+ *      VariableDeclaration: (node, typir) => { return [...]; },
+ *   });
+ * ```
+ *
+ * @param T a type definition mapping language specific type names (keys) to the corresponding types (values)
+ */
+export type LangiumValidationRules<T extends LangiumAstTypes> = {
+    [K in keyof T]?: T[K] extends AstNode ? ValidationRule<T[K]> | Array<ValidationRule<T[K]>> : never
+} & {
+    AstNode?: ValidationRule<AstNode> | Array<ValidationRule<AstNode>>;
+}
+
+
+export interface LangiumValidationCollector<RootType extends AstNode = AstNode> extends ValidationCollector<AstNode, RootType> {
+    addValidationRulesForAstNodes<AstTypes extends LangiumAstTypes>(rules: LangiumValidationRules<AstTypes>): void;
+}
+
+export class DefaultLangiumValidationCollector extends DefaultValidationCollector implements LangiumValidationCollector {
+
+    addValidationRulesForAstNodes<AstTypes extends LangiumAstTypes>(rules: LangiumValidationRules<AstTypes>): void {
+        // map this approach for registering validation rules to the key-value approach from core Typir
+        for (const [type, ruleCallbacks] of Object.entries(rules)) {
+            const languageKey = type === 'AstNode' ? undefined : type; // using 'AstNode' as key is equivalent to specifying no key
+            const callbacks = ruleCallbacks as ValidationRule | ValidationRule[];
+            if (Array.isArray(callbacks)) {
+                for (const callback of callbacks) {
+                    this.addValidationRule(callback, { languageKey });
+                }
+            } else {
+                this.addValidationRule(callbacks, { languageKey });
+            }
+        }
+    }
+
 }